Fix cppcheck issues (#1028)
Rename method parameters to match declaration and definition
Remove unused variables

Resolves: OLPEDGE-1062

+++ olp-cpp-sdk-core/src/cache/DefaultCacheImpl.cpp
@@ -565,7 +565,7 @@ int64_t DefaultCacheImpl::MaybeUpdatedProtectedKeys(
     } else if (prev_size > 0) {
       // delete key, as protected list is empty
       batch.Delete(kProtectedKeys);
-      return -1 * (prev_size + strlen(kProtectedKeys));
+      return -1 * static_cast<int64_t>((prev_size + strlen(kProtectedKeys)));
     }
   }
